Updating feature to add vrm metadata
Export VRM function updated as we need to change data of the GLB. As I could not find a way to edit the GLB directly. I am modifying the GLTF by adding VRM metadata for specification 1.0, due to I did not found examples of VRMs in spec 0.0 and tests was not working at all. This let me change VRM metadata properties for humanoid feature as we need to set the index of glTF.Node in each human bone.