Remove SingleMix endpoint, adapt calling function
The endpoint '/api/v1/singleMix' has been removed from the api.enum file as it is no longer in use. The function GetAvatarCombinationByAttributes in api.util file, which used this endpoint, has now been adapted to use the '/api/v1/single' endpoint. A POST request is now performed instead of a GET request. This change was necessary to align with the updates in the API specifications.